test code refactor
