+++ tests/unit/app/random-integer.test.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const randomInteger = require('../../app/random-integer.js');
+const randomInteger = require('../../../app/random-integer.js');
 
 test('Should return 0 when min and max is set to 0', () => {
   expect(randomInteger(0, 0)).toBe(0);
